added add employee functionality

+++ Helpers/Reporter.cs
@@ -25,9 +25,10 @@ public class Reporter{
          foreach(var programmerInCharge in projectTeam.ProgrammersInCharge){
             Console.WriteLine("- {0} {1} - In charge of {2} from {3} to {4} (duration: {5} days.), has worked a total of {7} hours, costing a total of {6}$", 
                programmerInCharge.LastName, programmerInCharge.FirstName, programmerInCharge.Activity, 
-               programmerInCharge.StartDate.ToString("MM/dd/yyyy"), programmerInCharge.EndDate.ToString("MM/dd/yyyy"), programmerInCharge.GetInterval().TotalDays, programmerInCharge.TotalCost(), programmerInCharge.WorkedDays); 
+               programmerInCharge.StartDate.ToString("MM/dd/yyyy"), programmerInCharge.EndDate.ToString("MM/dd/yyyy"), ((int)programmerInCharge.GetInterval().TotalDays), programmerInCharge.TotalCost(), programmerInCharge.WorkedDays); 
 
          }
       }
+      
    }
 }
